requirements.txt
python-telegram-bot
google-genai
Pillow
requests